# src/main/java/io/github/ingridbromer/quarkussocial/rest/dto/FollowersPerUserResponse.java
package io.github.ingridbromer.quarkussocial.rest.dto;

import lombok.Data;

import java.util.List;

@Data
public class FollowersPerUserResponse {
    private Integer followersCount;
    private List<FollowerResponse> content;
}