# src/main/java/io/github/ingridbromer/quarkussocial/rest/dto/PostResponse.java
package io.github.ingridbromer.quarkussocial.rest.dto;

import io.github.ingridbromer.quarkussocial.domain.model.Post;
import lombok.Data;

import java.time.LocalDateTime;

@Data
public class PostResponse {
    private String text;
    private LocalDateTime dateTime;

    public static PostResponse fromEntity(Post post){
        var response = new PostResponse();
        response.setText(post.getText());
        response.setDateTime(post.getDateTime());
        return response;
    }
}